Fix clippy: cast signal handler through usize instead of direct fn-to-int (#1563)
Newer clippy versions flag direct function-to-integer casts. Cast
through usize first to satisfy the lint.

+++ crates/cli/src/main.rs
@@ -480,8 +480,8 @@ fn ctrlc_handler(running: std::sync::Arc<std::sync::atomic::AtomicBool>) {
     let ptr = std::sync::Arc::into_raw(running) as *mut std::sync::atomic::AtomicBool;
     RUNNING_PTR.store(ptr, std::sync::atomic::Ordering::SeqCst);
     unsafe {
-        libc::signal(libc::SIGTERM, handle_signal as libc::sighandler_t);
-        libc::signal(libc::SIGINT, handle_signal as libc::sighandler_t);
+        libc::signal(libc::SIGTERM, handle_signal as usize as libc::sighandler_t);
+        libc::signal(libc::SIGINT, handle_signal as usize as libc::sighandler_t);
     }
 }
